Drop support for GHC 7.0 and 7.2
`call-stack`'s `HasCallStack` compatibility shim only supports back to GHC 7.4.
As a result, we can't reasonably support GHC 7.0 and 7.2 in `exceptions`
without a fair bit of grimy CPP. That being said, GHC 7.0 and 7.2 are quite
ancient by this point, so I'm comfortable with simply dropping support for
them. This patch makes the necessary `.cabal` and CI changes to accomplish that.