test: `cabal test -w ghc-9.4 --constraint="mtl == 2.3.1"`
The test with the following command did not pass because the decision to import the mtl module was made by referring to the transformers version.

~~~console
cabal test -w ghc-9.4 --constraint="mtl == 2.3.1"
~~~

There were two avenues to resolve this.
One is by making it refer to the mtl version,
The other is to change the transformers module to import.

It seems more natural to reference the mtl version, but since `mtl-2.3.1` still supports `transformers-0.5.6.2` which still provides `ErrorT` and so on, we can use `cabal test -w ghc-9.6 --constraint=" transformers=0.5.6.2" --constraint="mtl==2.3.1"`, the API will change when assuming `mtl-2.3.1`.
Therefore, we decided to change to import the transformers module.